+++ src/Symfony/Component/Cache/Tests/Simple/CacheTestCase.php
@@ -132,14 +132,9 @@ abstract class CacheTestCase extends SimpleCacheTest
     }
 }
 
-class NotUnserializable implements \Serializable
+class NotUnserializable
 {
-    public function serialize()
-    {
-        return serialize(123);
-    }
-
-    public function unserialize($ser)
+    public function __wakeup()
     {
         throw new \Exception(__CLASS__);
     }