+++ src/Symfony/Component/HttpKernel/Tests/KernelTest.php
@@ -330,9 +330,8 @@ EOF;
         $env = 'test_env';
         $debug = true;
         $kernel = new KernelForTest($env, $debug);
-
-        $expected = serialize([$env, $debug]);
-        $this->assertEquals($expected, $kernel->serialize());
+        $expected = "O:57:\"Symfony\Component\HttpKernel\Tests\Fixtures\KernelForTest\":2:{s:14:\"\0*\0environment\";s:8:\"test_env\";s:8:\"\0*\0debug\";b:1;}";
+        $this->assertEquals($expected, serialize($kernel));
     }
 
     /**