+++ src/Symfony/Component/Routing/CompiledRoute.php
@@ -50,7 +50,7 @@ class CompiledRoute implements \Serializable
     }
 
     /**
-     * {@inheritdoc}
+     * @internal since Symfony 4.3, will be removed in Symfony 5 as the class won't implement Serializable anymore
      */
     public function serialize()
     {
@@ -67,7 +67,7 @@ class CompiledRoute implements \Serializable
     }
 
     /**
-     * {@inheritdoc}
+     * @internal since Symfony 4.3, will be removed in Symfony 5 as the class won't implement Serializable anymore
      */
     public function unserialize($serialized)
     {