+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
  * added `block_prefix` option to `BaseType`.
  * added `help_html` option to display the `help` text as HTML.
  * `FormError` doesn't implement `Serializable` anymore
+ * `FormDataCollector` has been marked as `final`
  * added `label_translation_parameters`, `attr_translation_parameters`, `help_translation_parameters` options 
    to `FormType` to pass translation parameters to form labels, attributes (`placeholder` and `title`) and help text respectively.
    The passed parameters will replace placeholders in translation messages.