+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -8,6 +8,8 @@ CHANGELOG
  * added `CompiledUrlGenerator` and `CompiledUrlGeneratorDumper`
  * deprecated `PhpGeneratorDumper` and `PhpMatcherDumper`
  * deprecated `generator_base_class`, `generator_cache_class`, `matcher_base_class` and `matcher_cache_class` router options
+ * deprecated implementing `Serializable` for `Route` and `CompiledRoute`; if you serialize them, please
+   ensure your unserialization logic can recover from a failure related to an updated serialization format
 
 4.2.0
 -----